backend/requirements.txt
flask
flask-cors
flask-pymongo
groq
python-dotenv
pyjwt
bcrypt
certifi
